Allow polymorphism with `fine::ResourcePtr`
This feature should introduce no breaking change to the existing API,
while allowing resources to be subclasses and behave as such.

This requires a change to how `ResourceWrapper` works, but since it is
an implementation detail, this should not be obvious to users of the
library, causing no expected breaking changes.

The biggest downside of this feature is a slight increase in unused
padding inside the `ResourceWrapper` class. With our new approach, we
now add on average `alignof(std::max_align_t) - alignof(T)` unused
bytes per resource.